Bugfix Datentyp

+++ src/de/jost_net/JVerein/gui/view/AbstractAdresseDetailView.java
@@ -246,7 +246,7 @@ public abstract class AbstractAdresseDetailView extends AbstractView
 
       MitgliedDokument mido = (MitgliedDokument) Einstellungen.getDBService()
           .createObject(MitgliedDokument.class, null);
-      mido.setReferenz(new Integer(control.getMitglied().getID()));
+      mido.setReferenz(new Long(control.getMitglied().getID()));
       DokumentControl dcontrol = new DokumentControl(this, "mitglieder", true);
       cont.addPart(dcontrol.getDokumenteList(mido));
       ButtonArea butts = new ButtonArea();
